Remove functionality from script.js
Remove the functions under form.submit and windows.Domcontentloaded
and encapsulate them in Component class and Books class
